refactor(web-ui): extract UI components into separate files
