+++ src/web/routes/jobs/new.tsx
@@ -2,193 +2,9 @@ import type { FastifyInstance, FastifyRequest } from "fastify";
 import type { ScrapeTool } from "../../../tools/ScrapeTool"; // Adjusted import path
 import { ScrapeMode } from "../../../scraper/types"; // Adjusted import path
 import { logger } from "../../../utils/logger"; // Adjusted import path
-
-// --- Scrape Form Content Component (The actual form fields) ---
-const ScrapeFormContent = () => (
-  <div class="mt-4 p-4 bg-white dark:bg-gray-800 rounded-lg shadow border border-gray-300 dark:border-gray-600">
-    <h3 class="text-xl font-semibold text-gray-900 dark:text-white mb-2">
-      Queue New Scrape Job
-    </h3>
-    <form
-      hx-post="/api/jobs/scrape"
-      hx-target="#job-response"
-      hx-swap="innerHTML"
-      class="space-y-2"
-    >
-      <div>
-        <label
-          for="url"
-          class="block text-sm font-medium text-gray-700 dark:text-gray-300"
-        >
-          URL
-        </label>
-        <input
-          type="url"
-          name="url"
-          id="url"
-          required
-          class="mt-0.5 block w-full px-2 py-1 border border-gray-300 dark:border-gray-600 rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm bg-white dark:bg-gray-700 text-gray-900 dark:text-white"
-        />
-      </div>
-      <div>
-        <label
-          for="library"
-          class="block text-sm font-medium text-gray-700 dark:text-gray-300"
-        >
-          Library Name
-        </label>
-        <input
-          type="text"
-          name="library"
-          id="library"
-          required
-          class="mt-0.5 block w-full px-2 py-1 border border-gray-300 dark:border-gray-600 rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm bg-white dark:bg-gray-700 text-gray-900 dark:text-white"
-        />
-      </div>
-      <div>
-        <label
-          for="version"
-          class="block text-sm font-medium text-gray-700 dark:text-gray-300"
-        >
-          Version (optional)
-        </label>
-        <input
-          type="text"
-          name="version"
-          id="version"
-          class="mt-0.5 block w-full max-w-sm px-2 py-1 border border-gray-300 dark:border-gray-600 rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm bg-white dark:bg-gray-700 text-gray-900 dark:text-white"
-        />
-      </div>
-
-      {/* Consider using Flowbite Accordion here */}
-      <details class="bg-gray-50 dark:bg-gray-700 p-2 rounded-md">
-        <summary class="cursor-pointer text-sm font-medium text-gray-600 dark:text-gray-400">
-          Advanced Options
-        </summary>
-        <div class="mt-2 space-y-2">
-          <div>
-            <label
-              for="maxPages"
-              class="block text-sm font-medium text-gray-700 dark:text-gray-300"
-            >
-              Max Pages
-            </label>
-            <input
-              type="number"
-              name="maxPages"
-              id="maxPages"
-              min="1"
-              placeholder="1000"
-              class="mt-0.5 block w-full max-w-sm px-2 py-1 border border-gray-300 dark:border-gray-600 rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm bg-white dark:bg-gray-700 text-gray-900 dark:text-white"
-            />
-          </div>
-          <div>
-            <label
-              for="maxDepth"
-              class="block text-sm font-medium text-gray-700 dark:text-gray-300"
-            >
-              Max Depth
-            </label>
-            <input
-              type="number"
-              name="maxDepth"
-              id="maxDepth"
-              min="0"
-              placeholder="3"
-              class="mt-0.5 block w-full max-w-sm px-2 py-1 border border-gray-300 dark:border-gray-600 rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm bg-white dark:bg-gray-700 text-gray-900 dark:text-white"
-            />
-          </div>
-          <div>
-            <label
-              for="scope"
-              class="block text-sm font-medium text-gray-700 dark:text-gray-300"
-            >
-              Scope
-            </label>
-            <select
-              name="scope"
-              id="scope"
-              class="mt-0.5 block w-full max-w-sm pl-2 pr-10 py-1 text-base border border-gray-300 dark:border-gray-600 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm rounded-md bg-white dark:bg-gray-700 text-gray-900 dark:text-white"
-            >
-              <option value="subpages" selected>
-                Subpages (Default)
-              </option>
-              <option value="hostname">Hostname</option>
-              <option value="domain">Domain</option>
-            </select>
-          </div>
-          <div>
-            <label
-              for="scrapeMode"
-              class="block text-sm font-medium text-gray-700 dark:text-gray-300"
-            >
-              Scrape Mode
-            </label>
-            <select
-              name="scrapeMode"
-              id="scrapeMode"
-              class="mt-0.5 block w-full max-w-sm pl-2 pr-10 py-1 text-base border border-gray-300 dark:border-gray-600 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm rounded-md bg-white dark:bg-gray-700 text-gray-900 dark:text-white"
-            >
-              <option value={ScrapeMode.Auto} selected>
-                Auto (Default)
-              </option>
-              <option value={ScrapeMode.Fetch}>Fetch</option>
-              <option value={ScrapeMode.Playwright}>Playwright</option>
-            </select>
-          </div>
-          <div class="flex items-center">
-            <input
-              id="followRedirects"
-              name="followRedirects"
-              type="checkbox"
-              checked
-              class="h-4 w-4 text-indigo-600 focus:ring-indigo-500 border-gray-300 dark:border-gray-600 rounded bg-white dark:bg-gray-700"
-            />
-            <label
-              for="followRedirects"
-              class="ml-1 block text-sm text-gray-900 dark:text-gray-300"
-            >
-              Follow Redirects
-            </label>
-          </div>
-          <div class="flex items-center">
-            <input
-              id="ignoreErrors"
-              name="ignoreErrors"
-              type="checkbox"
-              checked
-              class="h-4 w-4 text-indigo-600 focus:ring-indigo-500 border-gray-300 dark:border-gray-600 rounded bg-white dark:bg-gray-700"
-            />
-            <label
-              for="ignoreErrors"
-              class="ml-1 block text-sm text-gray-900 dark:text-gray-300"
-            >
-              Ignore Errors During Scraping
-            </label>
-          </div>
-        </div>
-      </details>
-
-      <div>
-        <button
-          type="submit"
-          class="w-full flex justify-center py-1.5 px-3 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
-        >
-          Queue Job
-        </button>
-      </div>
-    </form>
-    {/* Target div for HTMX response */}
-    <div id="job-response" class="mt-2 text-sm"></div>
-  </div>
-);
-
-// --- Scrape Form Wrapper Component (For initial load and OOB target) ---
-const ScrapeForm = () => (
-  <div id="scrape-form-container">
-    <ScrapeFormContent />
-  </div>
-);
+import ScrapeForm from "../../components/ScrapeForm"; // Import extracted component
+import Alert from "../../components/Alert"; // Import Alert component
+import ScrapeFormContent from "../../components/ScrapeFormContent"; // Import for OOB swap
 
 /**
  * Registers the API routes for creating new jobs.
@@ -230,27 +46,13 @@ export function registerNewJobRoutes(
         // Basic validation
         if (!body.url || !body.library) {
           reply.status(400);
-          // Return JSX for validation error using Flowbite Alert
+          // Use Alert component for validation error
           return (
-            <div
-              class="flex items-center p-4 mb-4 text-sm text-red-800 border border-red-300 rounded-lg bg-red-50 dark:bg-gray-800 dark:text-red-400 dark:border-red-800"
-              role="alert"
-            >
-              <svg
-                class="flex-shrink-0 inline w-4 h-4 me-3"
-                aria-hidden="true"
-                xmlns="http://www.w3.org/2000/svg"
-                fill="currentColor"
-                viewBox="0 0 20 20"
-              >
-                <path d="M10 .5a9.5 9.5 0 1 0 9.5 9.5A9.51 9.51 0 0 0 10 .5ZM9.5 4a1.5 1.5 0 1 1 0 3 1.5 1.5 0 0 1 0-3ZM12 15H8a1 1 0 0 1 0-2h1v-3h-1a1 1 0 0 1 0-2h2a1 1 0 0 1 1 1v4h1a1 1 0 0 1 0 2Z" />
-              </svg>
-              <span class="sr-only">Info</span>
-              <div>
-                <span class="font-medium">Validation Error:</span> URL and
-                Library Name are required.
-              </div>
-            </div>
+            <Alert
+              type="error"
+              title="Validation Error:"
+              message="URL and Library Name are required."
+            />
           );
         }
 
@@ -279,29 +81,19 @@ export function registerNewJobRoutes(
         const result = await scrapeTool.execute(scrapeOptions);
 
         if ("jobId" in result) {
-          // Success: Return JSX Fragment with message and OOB swap using Flowbite Alert
+          // Success: Use Alert component and OOB swap
           return (
             <>
               {/* Main target response */}
-              <div
-                class="flex items-center p-4 mb-4 text-sm text-green-800 border border-green-300 rounded-lg bg-green-50 dark:bg-gray-800 dark:text-green-400 dark:border-green-800"
-                role="alert"
-              >
-                <svg
-                  class="flex-shrink-0 inline w-4 h-4 me-3"
-                  aria-hidden="true"
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="currentColor"
-                  viewBox="0 0 20 20"
-                >
-                  <path d="M10 .5a9.5 9.5 0 1 0 9.5 9.5A9.51 9.51 0 0 0 10 .5Zm9.5 9.5A9.5 9.5 0 0 1 10 19a9.46 9.46 0 0 1-1.671-.14c-.165-.05-.3-.19-.42-.335l-.165-.165c-.19-.2-.3-.425-.3-.655A4.2 4.2 0 0 1 4.5 10a4.25 4.25 0 0 1 7.462-2.882l1.217 1.217a3.175 3.175 0 0 0 4.5.01l.106-.106a.934.934 0 0 0 .1-.36ZM10 11a1 1 0 1 0 0 2 1 1 0 0 0 0-2Z" />
-                </svg>
-                <span class="sr-only">Info</span>
-                <div>
-                  <span class="font-medium">Success:</span> Job queued
-                  successfully! ID: <span safe>{result.jobId}</span>
-                </div>
-              </div>
+              <Alert
+                type="success"
+                message={
+                  <>
+                    Job queued successfully! ID:{" "}
+                    <span safe>{result.jobId}</span>
+                  </>
+                }
+              />
               {/* OOB target response - contains only the inner form content */}
               <div id="scrape-form-container" hx-swap-oob="innerHTML">
                 <ScrapeFormContent />
@@ -311,54 +103,21 @@ export function registerNewJobRoutes(
         }
 
         // This case shouldn't happen with waitForCompletion: false, but handle defensively
-        // Return JSX for unexpected success using Flowbite Alert
+        // Use Alert component for unexpected success
         return (
-          <div
-            class="flex items-center p-4 mb-4 text-sm text-yellow-800 border border-yellow-300 rounded-lg bg-yellow-50 dark:bg-gray-800 dark:text-yellow-300 dark:border-yellow-800"
-            role="alert"
-          >
-            <svg
-              class="flex-shrink-0 inline w-4 h-4 me-3"
-              aria-hidden="true"
-              xmlns="http://www.w3.org/2000/svg"
-              fill="currentColor"
-              viewBox="0 0 20 20"
-            >
-              <path d="M10 .5a9.5 9.5 0 1 0 9.5 9.5A9.51 9.51 0 0 0 10 .5ZM9.5 4a1.5 1.5 0 1 1 0 3 1.5 1.5 0 0 1 0-3ZM12 15H8a1 1 0 0 1 0-2h1v-3h-1a1 1 0 0 1 0-2h2a1 1 0 0 1 1 1v4h1a1 1 0 0 1 0 2Z" />
-            </svg>
-            <span class="sr-only">Info</span>
-            <div>
-              <span class="font-medium">Warning:</span> Job finished
-              unexpectedly quickly.
-            </div>
-          </div>
+          <Alert type="warning" message="Job finished unexpectedly quickly." />
         );
       } catch (error) {
         const errorMessage =
           error instanceof Error ? error.message : "Unknown error";
         logger.error(`Scrape job submission failed: ${error}`);
         reply.status(500); // Keep status code for errors
-        // Return JSX for server error using Flowbite Alert
+        // Use Alert component for server error
         return (
-          <div
-            class="flex items-center p-4 mb-4 text-sm text-red-800 border border-red-300 rounded-lg bg-red-50 dark:bg-gray-800 dark:text-red-400 dark:border-red-800"
-            role="alert"
-          >
-            <svg
-              class="flex-shrink-0 inline w-4 h-4 me-3"
-              aria-hidden="true"
-              xmlns="http://www.w3.org/2000/svg"
-              fill="currentColor"
-              viewBox="0 0 20 20"
-            >
-              <path d="M10 .5a9.5 9.5 0 1 0 9.5 9.5A9.51 9.51 0 0 0 10 .5ZM9.5 4a1.5 1.5 0 1 1 0 3 1.5 1.5 0 0 1 0-3ZM12 15H8a1 1 0 0 1 0-2h1v-3h-1a1 1 0 0 1 0-2h2a1 1 0 0 1 1 1v4h1a1 1 0 0 1 0 2Z" />
-            </svg>
-            <span class="sr-only">Info</span>
-            <div>
-              <span class="font-medium">Error:</span> Failed to queue job:{" "}
-              <span>{errorMessage}</span>
-            </div>
-          </div>
+          <Alert
+            type="error"
+            message={<>Failed to queue job: {errorMessage}</>}
+          />
         );
       }
     }
